Potential fix for code scanning alert no. 31: Multiplication result converted to larger type

+++ src/pke/lib/keyswitch/keyswitch-bv.cpp
@@ -230,8 +230,8 @@ EvalKey<DCRTPoly> KeySwitchBV::KeySwitchGenInternal(const PrivateKey<DCRTPoly> o
     auto elementParams    = newp0.GetParams();
 
     if (digitSize > 0) {
-        av.reserve(sOld.GetNumOfElements() * digitSize);
-        bv.reserve(sOld.GetNumOfElements() * digitSize);
+        av.reserve(static_cast<size_t>(sOld.GetNumOfElements()) * static_cast<size_t>(digitSize));
+        bv.reserve(static_cast<size_t>(sOld.GetNumOfElements()) * static_cast<size_t>(digitSize));
         for (usint i = 0; i < sOld.GetNumOfElements(); i++) {
             for (auto&& sOldDecomposed : sOld.GetElementAtIndex(i).PowersOfBase(digitSize)) {
                 DCRTPoly filtered(elementParams, Format::EVALUATION, true);
